Struct liquidfun::box2d::dynamics::fixture::FixtureDef

pub struct FixtureDef {
    pub shape: *mut B2Shape,
    pub user_data: size_t,
    pub friction: Float32,
    pub restitution: Float32,
    pub density: Float32,
    pub is_sensor: bool,
    pub filter: Filter,
}

A fixture definition is used to create a fixture. This class defines an abstract fixture definition. You can reuse fixture definitions safely.

Fields

shape

The shape, this must be set. The shape will be cloned, so you can create the shape on the stack.

user_data

Use this to store application specific fixture data.

friction

The friction coefficient, usually in the range [0,1].

restitution

The restitution (elasticity) usually in the range [0,1].

density

The density, usually in kg/m2.

is_sensor

A sensor shape collects contact information but never generates a collision response.

filter

Contact filtering data.

Methods

impl FixtureDef

fn new(shape: &Shape) -> FixtureDef