Struct liquidfun::box2d::dynamics::joints::JointDef

pub struct JointDef {
    pub joint_type: JointType,
    pub user_data: size_t,
    pub body_a: Option<Body>,
    pub body_b: Option<Body>,
    pub collide_connected: bool,
}

Joint definitions are used to construct joints.

Fields

joint_type

user_data

Use this to attach application specific data to your joints.

body_a

The first attached body.

body_b

The second attached body.

collide_connected

Set this flag to true if the attached bodies should collide.