Module liquidfun::box2d::particle

Modules

particle_color

particle_group

particle_system

Structs

B2ParticleDef

ParticleDef
A particle definition holds all the data needed to construct a particle. You can safely re-use these definitions.

ParticleFlags
The particle type. Can be combined with the | operator.

Constants

BARRIER_PARTICLE
Prevents other particles from leaking.

COLOR_MIXING_PARTICLE
Mix color between contacting particles.

DESTRUCTION_LISTENER_PARTICLE
Call b2DestructionListener on destruction.

ELASTIC_PARTICLE
With restitution from deformation.

FIXTURE_CONTACT_FILTER_PARTICLE
Call b2ContactFilter when this particle interacts with rigid bodies.

FIXTURE_CONTACT_LISTENER_PARTICLE
Call b2ContactListener when this particle is about to interact with a rigid body or stops interacting with a rigid body. This results in an expensive operation compared to using b2_fixtureContactFilterParticle to detect collisions between particles.

PARTICLE_CONTACT_FILTER_PARTICLE
Call b2ContactFilter when this particle interacts with other particles.

PARTICLE_CONTACT_LISTENER_PARTICLE
Call b2ContactListener when this particle is about to interact with another particle or stops interacting with another particle. This results in an expensive operation compared to using b2_particleContactFilterParticle to detect collisions between particles.

POWDER_PARTICLE
Without isotropic pressure.

REACTIVE_PARTICLE
Makes pairs or triads with other particles.

REPULSIVE_PARTICLE
With high repulsive force.

SPRING_PARTICLE
With restitution from stretching.

STATIC_PRESSURE_PARTICLE
Less compressibility.

TENSILE_PARTICLE
With surface tension.

VISCOUS_PARTICLE
With viscosity.

WALL_PARTICLE
Zero velocity.

WATER_PARTICLE
Water particle.

ZOMBIE_PARTICLE
Removed after next simulation step.